Added questions and working random button
