+++ query-quest/src/CategoryCard.js
@@ -1,5 +1,5 @@
 function CategoryCard({category}) {
- 
+    console.log(category)
     return (
      <div className="catTile ui card">
      <div>
