+++ query-quest/src/NavBar.js
@@ -6,7 +6,7 @@ function NavBar() {
   <div className="navbar">
     <nav>
       <NavLink exact to="/">Home  </NavLink>
-      <NavLink to="/movies">Play Game  </NavLink>
+      <NavLink to="/playgame">Play Game  </NavLink>
       <NavLink to="/newQuestion">New Question  </NavLink>
     </nav>  
   </div>)
